iTeam/requirements.txt
django>=1.6,<1.7
markdown>=2.1
django-email-obfuscator
crispy-forms
crispy-forms-foundation

requests
facebook-sdk

django-scheduler
